refactor: i18n refactoring
