+++ src/core/i18n/index.tsx
@@ -1,4 +1,4 @@
-import { locale } from 'expo-localization';
+import { getLocales } from 'expo-localization';
 import i18n from 'i18next';
 import { initReactI18next } from 'react-i18next';
 import { I18nManager } from 'react-native';
@@ -9,7 +9,7 @@ export * from './utils';
 
 i18n.use(initReactI18next).init({
   resources,
-  lng: getLanguage() || locale,
+  lng: getLanguage() || getLocales()[0].languageCode,
   fallbackLng: 'en',
   compatibilityJSON: 'v3', // By default React Native projects does not support Intl
