+++ src/core/i18n/resources.ts
@@ -9,3 +9,5 @@ export const resources = {
     translation: ar,
   },
 };
+
+export type Language = keyof typeof resources;
