+++ src/core/i18n/types.ts
@@ -1,5 +1,3 @@
-export type Language = 'en' | 'ar';
-
 //  https://github.com/infinitered/ignite/blob/master/boilerplate/app/i18n/i18n.ts
 export type RecursiveKeyOf<TObj extends object> = {
   [TKey in keyof TObj & (string | number)]: RecursiveKeyOfHandleValue<
